Move complex_walk to utils
